refactor(process): migrate expvars component to v2
Migrates `comp/process/expvars` from V1 to V2 component structure:
- `def/component.go` — interface only
- `impl/expvars.go` — implementation using compdef
- `fx/fx.go` — FX module

Updates all callers to use new import paths.
