+++ .golangci.yml
@@ -171,7 +171,7 @@ linters:
             - "!**/comp/process/agent/agent_linux.go"
             - "!**/comp/process/agent/status.go"
             - "!**/comp/process/apiserver/apiserver.go"
-            - "!**/comp/process/expvars/expvarsimpl/expvars.go"
+            - "!**/comp/process/expvars/impl/expvars.go"
             - "!**/comp/process/forwarders/forwardersimpl/forwarders.go"
             - "!**/comp/process/profiler/impl/profiler.go"
             - "!**/comp/process/status/statusimpl/status.go"
