+++ comp/process/expvars/impl/expvars.go
@@ -12,36 +12,28 @@ import (
 	"fmt"
 	"net/http"
 
-	"go.uber.org/fx"
-
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
 	"github.com/DataDog/datadog-agent/comp/core/telemetry"
-	"github.com/DataDog/datadog-agent/comp/process/expvars"
+	compdef "github.com/DataDog/datadog-agent/comp/def"
+	expvars "github.com/DataDog/datadog-agent/comp/process/expvars/def"
 	"github.com/DataDog/datadog-agent/comp/process/hostinfo/def"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/process/runner/endpoint"
 	"github.com/DataDog/datadog-agent/pkg/process/status"
 	"github.com/DataDog/datadog-agent/pkg/process/util"
 	sysconfig "github.com/DataDog/datadog-agent/pkg/system-probe/config"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newExpvarServer))
-}
 
 var _ expvars.Component = (*expvarServer)(nil)
 
 type expvarServer *http.Server
 
 type dependencies struct {
-	fx.In
+	compdef.In
 
-	Lc fx.Lifecycle
+	Lc compdef.Lifecycle
 
 	Config         config.Component
 	SysProbeConfig sysprobeconfig.Component
@@ -51,7 +43,8 @@ type dependencies struct {
 	Telemetry telemetry.Component
 }
 
-func newExpvarServer(deps dependencies) (expvars.Component, error) {
+// NewComponent creates a new expvars component.
+func NewComponent(deps dependencies) (expvars.Component, error) {
 	// Initialize status
 	err := InitProcessStatus(deps.Config, deps.SysProbeConfig, deps.HostInfo, deps.Log)
 	if err != nil {
@@ -67,7 +60,7 @@ func newExpvarServer(deps dependencies) (expvars.Component, error) {
 	expvarPort := getExpvarPort(deps)
 	expvarServer := &http.Server{Addr: fmt.Sprintf("localhost:%d", expvarPort), Handler: http.DefaultServeMux}
 
-	deps.Lc.Append(fx.Hook{
+	deps.Lc.Append(compdef.Hook{
 		OnStart: func(_ context.Context) error {
 			go func() {
 				err := expvarServer.ListenAndServe()
